backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Azure SDK
azure-storage-blob==12.19.0
azure-monitor-opentelemetry==1.2.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1